Move and rename fetch_and_store_artist_mbids.py
Move fetch_and_store_artist_mbids.py to db module
and rename it to artist.py
